+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Remove Bonemeal interactions between Fruit Leaves and Fruit Block
 - Make Fruit Leaves only valid bonemeal targets when there is air below the block
 - Make it so Flowering Leaves can't be bonemealed if placed by player
+- Give BWGFruitBlock a Codec
 
 # 2.2.2
 - Remove unneeded BWGPumpkin Class